Version 2.0 with AEFMs

+++ src/chmc-visualization/standard/standard.jl
@@ -46,77 +46,77 @@ pass through the initial state/node `I`. By default, these arrows stemming
 from the green nodes are omitted for visual clarity.
 """
 function plot_chmc(#
-  T::Matrix{<:Real},
-  I::Int64=1;
-  node_label_textsize::Real=15,
-  edge_label_textsize::Real=12,
-  arrow_shift::Real=0.85,
-  tfactor::Real=0.15,
-  x_pad::Real=0.75,
-  y_pad::Real=0.75,
-  tangents::Tuple{Tuple{<:Real, <:Real}, Tuple{<:Real, <:Real}} = ((1,0),(0,1)),
-  show_all::Bool=false
+    T::Matrix{<:Real},
+    I::Int64=1;
+    node_label_textsize::Real=15,
+    edge_label_textsize::Real=12,
+    arrow_shift::Real=0.85,
+    tfactor::Real=0.15,
+    x_pad::Real=0.75,
+    y_pad::Real=0.75,
+    tangents::Tuple{Tuple{<:Real, <:Real}, Tuple{<:Real, <:Real}} = ((1,0),(0,1)),
+    show_all::Bool=false
 )
 
-  # Error checking
-  sanitize_transition(T, I)
+    # Error checking
+    sanitize_transition(T, I)
 
-  # New Markov chain transition probability matrix
-  A, E, t = trie_adj_at_root(T, show_all, I)
+    # New Markov chain transition probability matrix
+    A, E, t = trie_adj_at_root(T, show_all, Int16(I))
 
-  # Vectors of node and edge colours
-  node_color = color_nodes(t, T, I)
-  edge_color = color_edges(A, E)
-  node_labels = label_nodes(t)
-  edge_labels = label_edges(A, E)
+    # Vectors of node and edge colours
+    node_color = color_nodes(t, T, I)
+    edge_color = color_edges(A, E)
+    node_labels = label_nodes(t)
+    edge_labels = label_edges(A, E)
 
-  # Tangents specify curved edges for looped EFM transitions
-  tangents = tangent_edges(A, E, tangents)
+    # Tangents specify curved edges for looped EFM transitions
+    tangents = tangent_edges(A, E, tangents)
 
-  # Construct simple digraph
-  G = SimpleDiGraph(A)
+    # Construct simple digraph
+    G = SimpleDiGraph(A)
 
-  # Construct Makie plot
-  f, ax, p = graphplot(#
-    G,
-    nlabels = node_labels,
-    elabels = edge_labels,
-    labels_textsize = node_label_textsize,
-    elabels_textsize = edge_label_textsize,
-    arrow_show = true,
-    arrow_shift = arrow_shift,
-    node_color = node_color,
-    tangents = tangents,
-    tfactor = repeat([tfactor], length(tangents)),
-    edge_color = edge_color
-  )
+    # Construct Makie plot
+    f, ax, p = graphplot(#
+        G,
+        nlabels = node_labels,
+        elabels = edge_labels,
+        labels_textsize = node_label_textsize,
+        elabels_textsize = edge_label_textsize,
+        arrow_show = true,
+        arrow_shift = arrow_shift,
+        node_color = node_color,
+        tangents = tangents,
+        tfactor = repeat([tfactor], length(tangents)),
+        edge_color = edge_color
+    )
 
-  # Update current plot with custom tree layout
-  x1, x2, y1, y2 = buchheim_layout(p, T, 1)
+    # Update current plot with custom tree layout
+    x1, x2, y1, y2 = buchheim_layout(p, T, 1)
 
-  # Padding boundary box
-  x1 -= x_pad
-  x2 += x_pad
-  y1 += y_pad
-  y2 -= y_pad
+    # Padding boundary box
+    x1 -= x_pad
+    x2 += x_pad
+    y1 += y_pad
+    y2 -= y_pad
 
-  # Remove plot decorations and center axes on CHMC
-  hidedecorations!(ax)
-  hidespines!(ax)
-  ax.aspect = DataAspect()
-  autolimits!(ax)
-  limits!(ax, x1, x2, y2, y1)
+    # Remove plot decorations and center axes on CHMC
+    hidedecorations!(ax)
+    hidespines!(ax)
+    ax.aspect = DataAspect()
+    autolimits!(ax)
+    limits!(ax, x1, x2, y2, y1)
 
-  # Makie node interactivity
-  function node_drag_action(state, idx, event, axis)
-      p[:node_pos][][idx] = event.data
-      p[:node_pos][] = p[:node_pos][]
-  end
-  ndrag = NodeDragHandler(node_drag_action)
-  deregister_interaction!(ax, :rectanglezoom)
-  register_interaction!(ax, :ndrag, ndrag)
+    # Makie node interactivity
+    function node_drag_action(state, idx, event, axis)
+        p[:node_pos][][idx] = event.data
+        p[:node_pos][] = p[:node_pos][]
+    end
+    ndrag = NodeDragHandler(node_drag_action)
+    deregister_interaction!(ax, :rectanglezoom)
+    register_interaction!(ax, :ndrag, ndrag)
 
-  # Return the Makie plot
-  f
+    # Return the Makie plot
+    f
 end
